DataManager 클래스 생성하여 App 클래스의 책임 분리
